requirements.txt
pymongo>=4.5.0
google-genai>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
numpy>=2.4.4
gunicorn>=21.2.0
rank-bm25>=0.2.2
